meta de faturamento passa para integer

+++ src/comercial/models.py
@@ -114,8 +114,7 @@ class MetaEstoqueCor(models.Model):
 
 class MetaFaturamento(models.Model):
     data = models.DateField()
-    faturamento = models.DecimalField(
-        decimal_places=2, max_digits=10, default=0)
+    faturamento = models.IntegerField(default=0)
 
     def __str__(self):
         return '{} - R$ {:.2}'.format(
